+++ cosmos_server/src/events/netty/netty_events.rs
@@ -164,9 +164,9 @@ fn handle_events_system(
                             bincode::serialize(&ServerReliableMessages::PlanetCreate {
                                 entity,
                                 body: NettyRigidBody::new(velocity, transform),
-                                width: structure.chunks_width(),
-                                height: structure.chunks_height(),
-                                length: structure.chunks_length(),
+                                width: structure.chunks_width() as u32,
+                                height: structure.chunks_height() as u32,
+                                length: structure.chunks_length() as u32,
                             })
                             .unwrap(),
                         );
@@ -177,9 +177,9 @@ fn handle_events_system(
                             bincode::serialize(&ServerReliableMessages::ShipCreate {
                                 entity,
                                 body: NettyRigidBody::new(velocity, transform),
-                                width: structure.chunks_width(),
-                                height: structure.chunks_height(),
-                                length: structure.chunks_length(),
+                                width: structure.chunks_width() as u32,
+                                height: structure.chunks_height() as u32,
+                                length: structure.chunks_length() as u32,
                             })
                             .unwrap(),
                         );
